Split out WaitFor{Challenge,Response}Confirmation
