+++ src/wallet/components/WaitForChallengeConfirmation.tsx
